refractor: reuse read form function
